Update README.md
Clarifies specifics about skipping Docker in Rails 7 because Rails 7 now generates a production Dockerfile that makes using Docker/Docker Compose in local development challenging.  The steps included in this document are only sufficient for Rails & Docker Compose in a development environment for Rails 7. (In order to run as-is with the production dockerfile, more steps are necessary for the compose, including passing the master key, etc.)

Signed-off-by: Kate Donaldson <kate@katelovescode.com>